+++ api/ConnecteurAPIController.class.php
@@ -1,29 +1,22 @@
 <?php
 
+use Pastell\Service\Connecteur\ConnecteurActionService;
 use Pastell\Service\Droit\DroitService;
 
 class ConnecteurAPIController extends BaseAPIController
 {
 
     private $donneesFormulaireFactory;
-
     private $connecteurEntiteSQL;
-
     private $actionPossible;
-
     private $fluxEntiteSQL;
-
     private $actionExecutorFactory;
-
     private $connecteurFactory;
-
     private $connecteurDefinitionFiles;
-
     private $jobManager;
-
     private $entiteSQL;
-
     private $droitService;
+    private $connecteurActionService;
 
     public function __construct(
         DonneesFormulaireFactory $donneesFormulaireFactory,
@@ -35,7 +28,8 @@ class ConnecteurAPIController extends BaseAPIController
         ConnecteurDefinitionFiles $connecteurDefinitionFiles,
         JobManager $jobManager,
         EntiteSQL $entiteSQL,
-        DroitService $droitService
+        DroitService $droitService,
+        ConnecteurActionService $connecteurActionService
     ) {
         $this->donneesFormulaireFactory = $donneesFormulaireFactory;
         $this->connecteurEntiteSQL = $connecteurEntiteSQL;
@@ -47,6 +41,7 @@ class ConnecteurAPIController extends BaseAPIController
         $this->jobManager = $jobManager;
         $this->entiteSQL = $entiteSQL;
         $this->droitService = $droitService;
+        $this->connecteurActionService = $connecteurActionService;
     }
 
     private function verifExists($id_ce)
@@ -273,6 +268,15 @@ class ConnecteurAPIController extends BaseAPIController
         }
 
         $id_ce =  $this->connecteurEntiteSQL->addConnecteur($id_e, $id_connecteur, $connecteur_info['type'], $libelle);
+
+        $this->connecteurActionService->add(
+            $id_e,
+            $this->getUtilisateurId(),
+            $id_ce,
+            '',
+            ConnecteurActionService::ACTION_AJOUTE,
+            "Le connecteur $libelle a été créé via l'API"
+        );
 
         //TODO Ajouter une fonction pour lancer les actions autos sur le connecteur
         //$this->jobManager->setJobForConnecteur($id_ce,$action_name,"création du connecteur");